Rename example→examples, bump to 100k threads, fix paths, add fat-jar packaging
- Rename to VirtualThreadsNativeJvmBenchmark, module to examples
- Bump from 10k to 100k virtual threads
- Add sbt-assembly for fat-jar packaging
- Move README content into class scaladoc with verified paths
- Retain only fat-jar (JVM) and native-binary methods
